+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## [Unreleased] - ReleaseDate
 - Add egui-based, GQLite browser.
 - Add support for in-memory databases.
+- Add `exponent` (^) operator.
 
 ## [0.5.2] - 2025-07-26
 - New Ruby/Python bindings.